- Throw an exception when importing ticket and an unknown user is discovered and cant be imported on-demand - Prevent running rerun cache from clearing out existing audit cache - Dont import user pictures on when fast_import is enabled - Fix a cache key on users which could cause uneccessary api calls later - Dont disable FK/unique checks when importing users - Add import flag with_preinserted_cache to indicate cache already exists and not to re-download it

+++ app/src/Application/DeskPRO/Import/Importer/Step/Zendesk/TicketsCacheStep.php
@@ -55,6 +55,10 @@ class TicketsCacheStep extends AbstractZendeskStep
 
 	public function countPages()
 	{
+		if ($this->importer->getConfig('with_preinserted_cache')) {
+			return 0;
+		}
+
 		$this->db->replace('import_datastore', array(
 			'typename' => 'zd_tickets_cache_time',
 			'data' => time()
@@ -87,6 +91,10 @@ class TicketsCacheStep extends AbstractZendeskStep
 
 	public function run($batch = 1)
 	{
+		if ($this->importer->getConfig('with_preinserted_cache')) {
+			return;
+		}
+
 		if ($batch == 1) {
 			$this->db->executeUpdate("DELETE FROM import_datastore WHERE typename LIKE 'zd_tickets_cache.%'");
 		}
